backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
python-dotenv
openai
